v0.0.3
This release contains a few small changes, not required unless you just
want the latest version.

- Adds ore generation to The End and The Nether. Please note if
generation seems lacking in the Overworld, it'll be the same in The End
and Nether as the generation code is exactly the same.
- Removed some debug code as it's no longer needed
- Work on #12
- Fixes #10 - Gives the "welcome" message some color and flair. Also
adds a configuration option to disable the message if you find it
annoying.